requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.3
python-dateutil==2.9.0.post0
langgraph==0.2.67
langchain-core==0.3.30
litellm==1.55.12
pytest==8.3.4
pytest-asyncio==0.24.0
